navigator: move ControlsStoreImpl to stores/
Now that ControlsControllerImpl is gone (commit a65918f5), this file
only contains ControlsStoreImpl — a store, not a controller. Move it
out of controllers/ to sit next to its peers in stores/.
